refactor: stabilize core helper boundaries (#466)
## Context

`@ontrails/core` still exposed several root helpers whose source modules lived under `src/internal/*`, which made the public surface look accidental and weakened the internal-import guardrail.

## What Changed

- Moved blessed core helper modules out of `src/internal/*` into stable source homes.
- Updated the core barrel and first-party imports to use those stable modules.
- Added a public-barrel regression test so future root exports do not point back into `./internal/*`.
- Tightened Warden public-internal-deep-import coverage for the new shape.

## How To Test

- `bun run --cwd packages/core typecheck`
- `bun test packages/core`
- `bun run lint`
- Final stack-tip verification: `bun run typecheck`, `bun run test`, `bun run lint`, `bun run check`, `bun run publish:check`

## Risks / Rollout Notes

The public root API is intentionally preserved; this is a source-ownership cleanup, not a consumer import break.

## Release / Changeset

Includes a changeset for `@ontrails/core` and `@ontrails/warden` because package contents and Warden guardrails changed.

Closes: TRL-643
