+++ packages/core/README.md
@@ -150,6 +150,21 @@ The developer returns `Result.err(new NotFoundError(...))`. The framework maps i
 - **Trail factories** (`@ontrails/core/trails`) -- derive CRUD-shaped trail contracts from contours without re-authoring IDs, schemas, examples, or intents
 - **Redaction** (`@ontrails/core/redaction`) -- strip sensitive data before logging
 
+### Public helper boundaries
+
+The root package also exposes a few low-level contracts that other framework
+packages build on:
+
+- **Intrinsic tracing** -- `TraceRecord`, `TraceSink`, `TraceContext`, and the
+  sink registry helpers are the core-owned execution record shape shared by
+  `@ontrails/observe`, `@ontrails/tracing`, and adapters.
+- **Trails DB** -- `deriveTrailsDbPath`, `deriveTrailsDir`,
+  `ensureSubsystemSchema`, `openReadTrailsDb`, and `openWriteTrailsDb` are the
+  generic database primitive used by framework subsystems.
+- **Surface projection helpers** -- layer field projection, cross-batch
+  validation, late-bound signal references, and Zod default-wrapper stripping
+  are stable root exports for first-party surfaces, store helpers, and tests.
+
 See the [API Reference](../../docs/api-reference.md) for the full list.
 
 ## Migration: topo-store moved to `@ontrails/topographer`
